Update Selenium version and adjust related code
Bumped up the required Selenium version from 3.9.0 to 4.1.0 in requirements.txt and setup.py files. To accommodate this Selenium version upgrade in the middlewares.py file, the ChromeDriverManager import statement was commented out and the driver initialization was simplified. The change was made due to the inbuilt driver manager in the latest Selenium versions.